+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Change log
 ## 5.1.1
 ### Back-end
-* 
+* Add reconnection.attempts, reconnection.delay options to config - applicable for editor-server connection